+++ _scripts/build_npm.ts
@@ -1,8 +1,8 @@
-import { build } from "jsr:@deno/dnt@0.41.2";
+import { build } from "jsr:@deno/dnt@0.41.3";
 import { dirname, join } from "jsr:@std/path@1.0.8";
-import { ensureDir } from "jsr:@std/fs@1.0.9";
+import { ensureDir } from "jsr:@std/fs@1.0.13";
 
-await Deno.remove("npm", { recursive: true }).catch(() => {});
+await Deno.remove("_npm", { recursive: true }).catch(() => {});
 
 const version = Deno.args[0];
 
@@ -11,11 +11,13 @@ if (!version) {
 }
 
 const testFiles = [
+  "changelog.azdo.md",
   "changelog.custom.type.md",
-  "changelog.expected.md",
   "changelog.expected.linted.md",
-  "changelog.md",
+  "changelog.expected.md",
   "changelog.gitlab.md",
+  "changelog.sort.md",
+  "changelog.md",
   "empty.expected.md",
 ];
 
@@ -38,10 +40,11 @@ await build({
     {
       kind: "bin",
       name: "changelog",
-      path: "bin.ts",
+      path: "./bin.ts",
     },
   ],
-  outDir: "./npm",
+  compilerOptions: { target: "ES2022" },
+  outDir: "./_npm",
   package: {
     name: "keep-a-changelog",
     version,
@@ -66,7 +69,7 @@ await build({
 });
 
 async function copyFile(from: string, to = from) {
-  to = join("npm", to);
+  to = join("_npm", to);
   await ensureDir(dirname(to));
   await Deno.copyFile(from, to);
 }