+++ src/Changelog.ts
@@ -1,5 +1,5 @@
-import Release from "./Release.ts";
 import { equals, parse } from "./deps.ts";
+import type Release from "./Release.ts";
 import type { SemVer } from "./deps.ts";
 
 export default class Changelog {
@@ -13,24 +13,32 @@ export default class Changelog {
   tagNameBuilder?: (release: Release) => string;
   /** @deprecated: Use tagLinkBuilder() instead */
   compareLinkBuilder?: (previous: Release, release: Release) => string;
-  tagLinkBuilder?: (url: string, tag: string, previous?: string, head?: string) => string;
+  tagLinkBuilder?: (
+    url: string,
+    tag: string,
+    previous?: string,
+    head?: string,
+  ) => string;
   format: "compact" | "markdownlint" = "compact";
   bulletStyle: "-" | "*" | "+" = "-";
+  autoSortReleases = true;
 
   constructor(title: string, description = "") {
     this.title = title;
     this.description = description;
   }
 
-  addRelease(release: Release) {
+  addRelease(release: Release): this {
     this.releases.push(release);
-    this.sortReleases();
+    if (this.autoSortReleases) {
+      this.sortReleases();
+    }
     release.changelog = this;
 
     return this;
   }
 
-  findRelease(version?: SemVer | string) {
+  findRelease(version?: SemVer | string): Release | undefined {
     if (!version) {
       return this.releases.find((release) => !release.version);
     }
@@ -41,11 +49,11 @@ export default class Changelog {
     );
   }
 
-  sortReleases() {
+  sortReleases(): void {
     this.releases.sort((a, b) => a.compare(b));
   }
 
-  compareLink(previous: Release, release: Release) {
+  compareLink(previous: Release, release: Release): string {
     if (this.compareLinkBuilder) {
       return this.compareLinkBuilder(previous, release);
     }
@@ -54,11 +62,21 @@ export default class Changelog {
       const url = this.url!;
 
       if (!previous) {
-        return this.tagLinkBuilder(this.url!, this.tagName(release), undefined, this.head);
+        return this.tagLinkBuilder(
+          this.url!,
+          this.tagName(release),
+          undefined,
+          this.head,
+        );
       }
 
       if (!release.date || !release.version) {
-        return this.tagLinkBuilder(url, this.head, this.tagName(previous), this.head);
+        return this.tagLinkBuilder(
+          url,
+          this.head,
+          this.tagName(previous),
+          this.head,
+        );
       }
 
       return this.tagLinkBuilder(
@@ -82,7 +100,7 @@ export default class Changelog {
     }`;
   }
 
-  tagName(release: Release) {
+  tagName(release: Release): string {
     if (this.tagNameBuilder) {
       return this.tagNameBuilder(release);
     }
@@ -90,7 +108,7 @@ export default class Changelog {
     return `v${release.version}`;
   }
 
-  toString() {
+  toString(): string {
     const t = [];
 
     if (this.flag) {